% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{count_processing}
\alias{count_processing}
\title{The main sampling function.}
\usage{
count_processing(count, remaining, opt)
}
\arguments{
\item{opt}{A list of parameters in Gibbs script
remaining A vector from predict_gene
count A table include the sampling count of each round of iteration}
}
\value{
A talbe
count  A data.table including all posterior probability and region risk p.
}
\description{
The main sampling function.
}
\examples{


}
