% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evi_direction.R
\name{evi_direction}
\alias{evi_direction}
\title{Conver the direction of evidence according to their column names.}
\usage{
evi_direction(ori)
}
\arguments{
\item{ori}{A data.table including the evidence data and column names containing the directons greater/less.}
}
\value{
A data.table with correct directions (Greater is better)
}
\description{
Conver the direction of evidence according to their column names.
}
\examples{
evi <- evi_direction(evi) 

}
