% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_evi.R
\name{extra_evi}
\alias{extra_evi}
\title{Processing extra evidence.}
\usage{
extra_evi(opt)
}
\arguments{
\item{opt}{A list of parameters}
}
\value{
A list containing extra_processed and extra_weight used in next step
}
\description{
Processing extra evidence.
}
\examples{
opt$extra <- extra_evi(opt)
}
