% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_candidates.R
\name{extract_candidates}
\alias{extract_candidates}
\title{Extract candidates from snp regions.}
\usage{
extract_candidates(opt)
}
\arguments{
\item{opt}{A list of parameters in Gibbs}
}
\value{
A data.table containing the genes in snp regions
}
\description{
Extract candidates from snp regions.
}
\examples{
opt$gene<-extract_candidates(opt)
}
