% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_evi.R
\name{generic_evi}
\alias{generic_evi}
\title{Process generic evidence and generate generic weight.}
\usage{
generic_evi(opt)
}
\arguments{
\item{opt}{A list of parameters in Gibbs script}
}
\value{
A list containing generic_proccesed and generic_weight used in next step
}
\description{
Process generic evidence and generate generic weight.
}
\examples{
opt$generic <- generic_evi(opt)
}
