% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_dis.R
\name{joint_dis}
\alias{joint_dis}
\title{Return joint distribution of two genes in sampling.}
\usage{
joint_dis(opt)
}
\arguments{
\item{opt}{A list of parameters in Gibbs script}
}
\value{
A data.table of coocurrence frequence
}
\description{
Return joint distribution of two genes in sampling.
}
\examples{

}
