% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_network.R
\name{load_network}
\alias{load_network}
\title{Returning the network string.}
\usage{
load_network(opt)
}
\arguments{
\item{opt}{A list of parameters in Gibbs script}
}
\value{
A network string based on the parameter inputed in Gibbs.
}
\description{
Returning the network string.
}
\examples{
opt$transp <- load_network(opt)
}
