% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis_transformation.R
\name{mahalanobis_transformation}
\alias{mahalanobis_transformation}
\title{Mahalanobis transformation of evidence.}
\usage{
mahalanobis_transformation(dt)
}
\arguments{
\item{dt}{a data.table containig the evidence data, without row names.}
}
\value{
a data.table containing the decorred features
}
\description{
Mahalanobis transformation of evidence.
}
\examples{
weight <- mahalanobis_transformation(evi)

}
