% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_chosen_update.R
\name{region_chosen_update}
\alias{region_chosen_update}
\title{Whether a region is chosen as a conditional-on region in the next iteration. Combined the prior calculated based on a beta distribution of region pvalue and the network contribution.}
\usage{
region_chosen_update(pickup, region_chosen, remaining, opt)
}
\arguments{
\item{pickup}{A number indicate the region index.
region_chosen A vector of TRUE/FALSE. The chosen region status in the last iteration
remaining A vector of genes. The candidates gene from last iteration.
opt}
}
\value{
Logical. TRUE when the region is chosen.
}
\description{
Whether a region is chosen as a conditional-on region in the next iteration. Combined the prior calculated based on a beta distribution of region pvalue and the network contribution.
}
\examples{
 region_chosen<-unlist(foreach(pickup=1:length(opt$region)) \%dopar\% region_chosen_update(pickup,region_chosen,remaining,opt))

}
