% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampling}
\alias{sampling}
\title{The main sampling function.}
\usage{
sampling(opt_tmp, run_step, with_region)
}
\arguments{
\item{opt}{A list of parameters in Gibbs script
mode String. burnin/post_burnin
with_region Logical. Whether the region risk is evaluated}
}
\value{
A list.
count  A data.table including all posterior probability and region risk p.
chosen A data.table ncluding the genes chosen in each round.
region_risk A data.table including the TRUE/FALSE in each round.
}
\description{
The main sampling function.
}
\examples{
burnin <- sampling(opt,mode="burnin",with_region=FALSE)

}
